refactor: extract createCreditNoteForInvoiceButtonProps
